Process Profiles and Extensions with StructureDefinitionProcessor
Profiles and Extensions are similar enough that they can be processed by
the same class. The ProfileProcessor and ExtensionProcessor classes are
removed. Test cases are reorganized to more directly test functions of
StructureDefinitionProcessor.
