+++ src/processor/FHIRProcessor.ts
@@ -1,10 +1,7 @@
 import fs from 'fs-extra';
 import { fhirdefs } from 'fsh-sushi';
 import { logger } from '../utils';
-import { ProfileProcessor } from './ProfileProcessor';
-import { ExtensionProcessor } from './ExtensionProcessor';
-import { CodeSystemProcessor } from './CodeSystemProcessor';
-import { ConfigurationProcessor } from './ConfigurationProcessor';
+import { StructureDefinitionProcessor, CodeSystemProcessor, ConfigurationProcessor } from '.';
 import {
   ExportableProfile,
   ExportableExtension,
@@ -37,13 +34,8 @@ export class FHIRProcessor {
       // Profiles and Extensions are both made from StructureDefinitions
       // Invariants may be contained within StructureDefinitions
       this.structureDefinitions.push(rawContent);
-      if (rawContent.type === 'Extension') {
-        logger.debug(`Processing contents of ${inputPath} as Extension.`);
-        return ExtensionProcessor.process(rawContent, this.fhir, existingInvariants);
-      } else {
-        logger.debug(`Processing contents of ${inputPath} as Profile.`);
-        return ProfileProcessor.process(rawContent, this.fhir, existingInvariants);
-      }
+      logger.debug(`Processing contents of ${inputPath} as StructureDefinition.`);
+      return StructureDefinitionProcessor.process(rawContent, this.fhir, existingInvariants);
     } else if (rawContent['resourceType'] === 'CodeSystem') {
       logger.debug(`Processing contents of ${inputPath} as CodeSystem.`);
       return [CodeSystemProcessor.process(rawContent)];
