+++ src/processor/StructureDefinitionProcessor.ts
@@ -1,6 +1,11 @@
 import compact from 'lodash/compact';
 import { fhirdefs } from 'fsh-sushi';
-import { ExportableSdRule, ExportableInvariant } from '../exportable';
+import {
+  ExportableSdRule,
+  ExportableInvariant,
+  ExportableProfile,
+  ExportableExtension
+} from '../exportable';
 import {
   CardRuleExtractor,
   CaretValueRuleExtractor,
@@ -15,7 +20,34 @@ import {
 import { ProcessableElementDefinition } from '.';
 import { getAncestorElement } from '../utils';
 
-export abstract class AbstractSDProcessor {
+export class StructureDefinitionProcessor {
+  static process(
+    input: any,
+    fhir: fhirdefs.FHIRDefinitions,
+    existingInvariants: ExportableInvariant[] = []
+  ): [ExportableProfile | ExportableExtension, ...ExportableInvariant[]] | [] {
+    if (StructureDefinitionProcessor.isProcessableStructureDefinition(input)) {
+      let sd: ExportableProfile | ExportableExtension;
+      if (input.type === 'Extension') {
+        sd = new ExportableExtension(input.name);
+      } else {
+        sd = new ExportableProfile(input.name);
+      }
+      const elements =
+        input.differential?.element?.map(rawElement => {
+          return ProcessableElementDefinition.fromJSON(rawElement, false);
+        }) ?? [];
+      StructureDefinitionProcessor.extractKeywords(input, sd);
+      const invariants = StructureDefinitionProcessor.extractInvariants(
+        elements,
+        existingInvariants
+      );
+      StructureDefinitionProcessor.extractRules(input, elements, sd, fhir);
+      return [sd, ...invariants];
+    }
+    return [];
+  }
+
   static extractKeywords(input: ProcessableStructureDefinition, target: ConstrainableEntity): void {
     // Usually name is already set (by constructor), but it doesn't hurt to be sure
     target.name = input.name;
@@ -90,6 +122,7 @@ export abstract class AbstractSDProcessor {
 export interface ProcessableStructureDefinition {
   name: string;
   resourceType: string;
+  type?: string;
   id?: string;
   title?: string;
   description?: string;
