+++ src/processor/index.ts
@@ -1,8 +1,7 @@
 // processors are responsible for transforming FHIR types (ex. StructureDefinition) into FSH types (ex. Profile)
 export * from './Package';
 export * from './FHIRProcessor';
-export * from './ProfileProcessor';
-export * from './ExtensionProcessor';
 export * from './CodeSystemProcessor';
 export * from './ProcessableElementDefinition';
 export * from './StructureDefinitionProcessor';
+export * from './ConfigurationProcessor';
