+++ test/processor/FHIRProcessor.test.ts
@@ -1,40 +1,36 @@
 import path from 'path';
 import {
-  ProfileProcessor,
-  ExtensionProcessor,
   FHIRProcessor,
-  CodeSystemProcessor
+  CodeSystemProcessor,
+  StructureDefinitionProcessor
 } from '../../src/processor';
 import '../helpers/loggerSpy'; // suppresses console logging
 import { fhirdefs } from 'fsh-sushi';
 
 describe('FHIRProcessor', () => {
   let processor: FHIRProcessor;
-  let profileSpy: jest.SpyInstance;
-  let extensionSpy: jest.SpyInstance;
+  let structureDefinitionSpy: jest.SpyInstance;
   let codeSystemSpy: jest.SpyInstance;
 
   beforeAll(() => {
     processor = new FHIRProcessor(new fhirdefs.FHIRDefinitions());
-    profileSpy = jest.spyOn(ProfileProcessor, 'process');
-    extensionSpy = jest.spyOn(ExtensionProcessor, 'process');
+    structureDefinitionSpy = jest.spyOn(StructureDefinitionProcessor, 'process');
     codeSystemSpy = jest.spyOn(CodeSystemProcessor, 'process');
   });
 
   beforeEach(() => {
-    profileSpy.mockClear();
-    extensionSpy.mockClear();
+    structureDefinitionSpy.mockClear();
     codeSystemSpy.mockClear();
   });
 
-  it('should try to process a Profile with the ProfileProcessor', () => {
+  it('should try to process a Profile with the StructureDefinitionProcessor', () => {
     processor.process(path.join(__dirname, 'fixtures', 'simple-profile.json'));
-    expect(profileSpy).toHaveBeenCalledTimes(1);
+    expect(structureDefinitionSpy).toHaveBeenCalledTimes(1);
   });
 
-  it('should try to process an Extension with the ExtensionProcessor', () => {
+  it('should try to process an Extension with the StructureDefinitionProcessor', () => {
     processor.process(path.join(__dirname, 'fixtures', 'simple-extension.json'));
-    expect(extensionSpy).toHaveBeenCalledTimes(1);
+    expect(structureDefinitionSpy).toHaveBeenCalledTimes(1);
   });
 
   it('should try to process a CodeSystem with the CodeSystemProcessor', () => {
